+++ CHANGELOG.md
@@ -12,6 +12,15 @@
 # 5.0.101
 - **bug-fixed**: removed private note msgs
 
+# 5.1.0-rc18
+- **bug-fixed**: allowedOnSpecificUrl 
+
+# 5.1.0-rc17
+- **added**: allowedOnSpecificUrl and allowedOnSpecificUrlList to manage list of pattern url to allow to load the widget
+
+# 5.1.0-rc16
+- **added**: hideOnSpecificUrlList replaced in favour of hideOnSpecificDomainList
+
 # 5.1.0-rc15
 # 5.1.0-rc14
 - **added**: ability to hide widget on selected domains based on hideOnSpecificDomainList variable